Testing framework for lshw
First stab at a testing framework for lshw using `chroot`. It uses
fakechroot/fakeroot so that no root permission is required to run the
tests.

It is also limited to testing only the CPU information. The idea is to
test only those data which are retrieved exclusively by reading from files in
/proc and /sys (for example) and hence testable without needing to
mock/patch any specific library/system call.

Change-Id: I0b4bdbb408fcc27087ddc30c8b1c17c010dbd771