devicetree: use defined sizes for integer property values (#690)
Integer properties in the device tree are defined in terms of "cells",
which are 32-bit integers in big-endian byte order. Most integer
properties are encoded as a single cell. The format of the "reg"
property is a bit more complicated, the lengths of the address and size
depends on the #address-cells and #size-cells properties (respectively)
on the parent node.

The code was previously using long for reading integer values, but in
some cases this is the wrong size (for example, on ppc64) causing many
properties to not be read.

On the other hand, it was using uint64_t for reading the "reg" property
which is the wrong size on 32-bit platforms.

This patch fixes the device tree parsing code to always use proper
sizes, regardless of the size of the platform types.

This also makes it possible to run cross-platform unit tests
(e.g. parsing a device tree on x86_64) since the code will no longer
assume that the endianness and integer sizes of the platform match the
device tree being read.