use sysfs to find businfo for SCSI adapters (#692)
In spite of its name, the SCSI_IOCTL_GET_PCI ioctl will actually return
the kernel name for the underlying SCSI adapter no matter whether it's
on the PCI bus or not. For example an IBM VIO SCSI adapter on the vio
bus might return 30000002 for this ioctl.

Since this doesn't look like a PCI bus address, the guessBusInfo()
function would then make a bad guess, returning businfo as
parisc@30000002 which is not right.

We can figure out the right businfo without guessing by looking in sysfs
instead.